[package][mediacenter-addon-osmc][script.module.osmcsetting.remotes] v0.1.0
