+++ package/mediacenter-addon-osmc/src/script.module.osmcsetting.remotes/addon.xml
@@ -1,23 +1,22 @@
 <?xml version="1.0" encoding="utf-8" standalone="yes"?>
-<addon id="script.module.osmcsetting.remotes"
-       name="Remotes"
-       version="0.0.1"
-       provider-name="OSMC (KodeKarnage)">
+<addon id="script.module.osmcsetting.remotes" name="Remotes" version="0.1.0" provider-name="OSMC (KodeKarnage)">
     <requires>
-        <import addon="xbmc.python" version="2.1.0"/>
+        <import addon="xbmc.python" version="2.26.0"/>
+        <import addon="script.module.osmccommon" version="0.1.0"/>
     </requires>
 
-<!-- <extension point="xbmc.python.script" library="default.py"/> -->
-<extension point="xbmc.python.module" library="lib" />
+    <extension point="xbmc.python.module" library="resources/lib/"/>
 
     <extension point="xbmc.addon.metadata">
-        <summary lang="en">osmcsetting.remotes</summary>
-        <description lang="en"></description>
+        <description lang="en_GB">My OSMC settings module.</description>
         <language>en</language>
         <platform>all</platform>
-        <license>GNU GPLv2</license>
+        <license>GPL-2.0-or-later</license>
         <email>karnage@osmc.tv</email>
-        <forum></forum>
-        <source></source>
+        <source>https://osmc.tv</source>
+        <assets>
+            <icon>resources/media/icon.png</icon>
+            <fanart>resources/media/fanart.jpg</fanart>
+        </assets>
     </extension>
 </addon>
