+++ package/mediacenter-addon-osmc/src/script.module.osmcsetting.remotes/resources/lib/osmcremotes/__init__.py
@@ -8,4 +8,4 @@
     See LICENSES/GPL-2.0-or-later for more information.
 """
 
-__all__ = ['osmcremotes']
+__all__ = ['remote_gui', 'remote_remotes', 'osmc']
