+++ package/mediacenter-addon-osmc/src/script.module.osmcsetting.remotes/resources/media/icon.png
