+++ package/mediacenter-addon-osmc/src/script.module.osmcsetting.remotes/resources/settings.xml
@@ -1,7 +1,5 @@
 <?xml version="1.0" encoding="utf-8" standalone="yes"?>
 <settings>
-	<category label="">
 
-	</category>
 </settings>
