+++ package/mediacenter-addon-osmc/src/script.module.osmcsetting.remotes/resources/skins/Default/1080i/OSMC_remote_testing720.xml
@@ -14,18 +14,6 @@
             <animation effect="fade" end="100" start="0" time="200">WindowOpen</animation>
             <animation effect="fade" end="0" start="100" time="200">WindowClose</animation>
         </control>
-        <!-- countdown label -->
-        <!--         <control id="45" type="label">-->
-        <!-- <left>40</left>-->
-        <!-- <top>262</top>-->
-        <!-- <width>270</width>-->
-        <!-- <height>36</height>-->
-        <!-- <align>center</align>-->
-        <!-- <label>nothing</label>-->
-        <!-- <font>osmc_addon_720_XLarge</font>-->
-        <!--former_line            <font>osmc_addon_720_OLD_Font48</font>-->
-        <!-- </control>-->
-        <!-- progress bar -->
         <control id="101" type="image">
             <left>360</left>
             <top>200</top>
@@ -64,9 +52,7 @@
                 <height>48</height>
                 <align>center</align>
                 <font>osmc_addon_720_XLarge</font>
-                <!--former_line<font>osmc_addon_720_OLD_Font48</font>-->
                 <textcolor>ff808080</textcolor>
-                <!--former_line<textcolor>ffffffff</textcolor>-->
             </control>
             <!-- restarting service label -->
             <control id="90" type="label">
@@ -77,7 +63,6 @@
                 <height>158</height>
                 <align>center</align>
                 <font>osmc_addon_720_Large</font>
-                <!--former_line<font>osmc_addon_720_OLD_Font42</font>-->
                 <textcolor>66ffffff</textcolor>
             </control>
             <!-- service restarted label -->
@@ -89,7 +74,6 @@
                 <height>158</height>
                 <align>center</align>
                 <font>osmc_addon_720_Large</font>
-                <!--former_line<font>osmc_addon_720_OLD_Font42</font>-->
                 <textcolor>66ffffff</textcolor>
             </control>
             <!-- No -->
@@ -107,7 +91,6 @@
                 <texturenofocus>single_pixel.png</texturenofocus>
                 <focusedcolor>FFFFFFFF</focusedcolor>
                 <textcolor>66ffffff</textcolor>
-                <!--former_line<font>osmc_addon_720_OLD_Font36</font>-->
             </control>
             <!-- Yes -->
             <control id="25" type="button">
@@ -124,7 +107,6 @@
                 <texturenofocus>single_pixel.png</texturenofocus>
                 <focusedcolor>FFFFFFFF</focusedcolor>
                 <textcolor>66ffffff</textcolor>
-                <!--former_line<font>osmc_addon_720_OLD_Font36</font>-->
             </control>
         </control>
     </controls>
